+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fixed a crash on variables that are set to strings.
 * Fixed an issue inserting values into an empty list.
 * Fixed various issues with viewing values set to symbols.
+* Added the ability to cycle values or list elements within custom
+  variables.
 
 # v0.1